+++ liberty-starter-application/skeletons/specialFiles/GitHubReadme.md
@@ -1,0 +1,7 @@
+# %NAME%
+
+This project was generated by the [Liberty app accelerator](%GENERATED_FROM%). To build the code clone the repository and run:
+
+```
+%RUN_INSTRUCTION%
+```
