+++ liberty-starter-application/src/test/java/com/ibm/liberty/starter/it/api/v1/RunLocalMavenPayloadTest.java
@@ -26,8 +26,9 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.PrintStream;
 
-import static com.ibm.liberty.starter.it.api.v1.matchers.FileContainsLines.containsLinesInRelativeOrder;
-import static com.ibm.liberty.starter.it.api.v1.matchers.Retry.eventually;
+import static com.ibm.liberty.starter.matchers.FileContainsLines.containsLinesInRelativeOrder;
+import static com.ibm.liberty.starter.matchers.FileContainsLines.containsLinesInRelativeOrder;
+import static com.ibm.liberty.starter.matchers.Retry.eventually;
 import static org.hamcrest.Matchers.containsString;
 import static org.hamcrest.Matchers.is;
 import static org.hamcrest.io.FileMatchers.anExistingFile;
